+++ addons/yat/src/scenes/monitor/components/memory_info/MemoryInfo.cs
@@ -29,13 +29,13 @@ public partial class MemoryInfo : PanelContainer, IMonitorComponent
 
 		_label.Clear();
 		_label.AppendText(
-			$"RAM: {NumericHelper.SizeToString(physical.AsInt64(), 3)}\n" +
+			$"RAM: {Numeric.SizeToString(physical.AsInt64(), 3)}\n" +
 			$"Free: {(
 				freePercent < 15 && UseColors
-				? $"[color={_yat.OptionsManager.Options.ErrorColor}]{NumericHelper.SizeToString(free.AsInt64(), 3)}[/color]"
-				: NumericHelper.SizeToString(free.AsInt64(), 3))}\n" +
-			$"Stack: {NumericHelper.SizeToString(stack.AsInt64(), 1)}\n" +
-			$"VRAM: {NumericHelper.SizeToString((long)vram, 2)}"
+				? $"[color={_yat.OptionsManager.Options.ErrorColor}]{Numeric.SizeToString(free.AsInt64(), 3)}[/color]"
+				: Numeric.SizeToString(free.AsInt64(), 3))}\n" +
+			$"Stack: {Numeric.SizeToString(stack.AsInt64(), 1)}\n" +
+			$"VRAM: {Numeric.SizeToString((long)vram, 2)}"
 		);
 	}
 }